--- MultiplesOfThreeOrFive/README.md
@@ -1,13 +1,0 @@
-# Multiples of 3 or 5
-## Problem 1
-
-If we list all the natural numbers below 10 that are multiples of 3 or 5, we get3, 5, 6 and 9. The sum of these multiples is 23.
-
-Find the sum of all the multiples of 3 or 5 below 1000.
-
-https://projecteuler.net/problem=1
-
-### Answer
-```
-233,168
-```
