--- Validator/README.md
@@ -1,3 +1,0 @@
-# Validator
-
-This is the shared functionality for validating data types for pre and post conditions
