+++ even_fibonacci.py
@@ -1,4 +1,4 @@
-from Validator.implementation import Validator
+from validator import Validator
 
 
 class Fibonacci:
@@ -7,6 +7,13 @@ class Fibonacci:
     
     def get_even_sum(self, limit):
         """
+        Even Fibonacci Numbers
+        Problem 2
+
+        Each new term in the Fibonacci sequence is generated by adding the previous two terms. By starting with $1$ and $2$, the first $10$ terms will be:
+        `1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...`
+        By considering the terms in the Fibonacci sequence whose values do not exceed four million, find the sum of the even-valued terms.
+
         precondition:
         limit is a natural number.
