+++ multiples_of_three_or_five.py
@@ -1,5 +1,5 @@
 
-from Validator.implementation import Validator
+from validator import Validator
 
 
 class Multiples:
