+++ test/test_even_fibonacci.py
@@ -1,14 +1,14 @@
 import unittest
-from . import implementation
+from even_fibonacci import Fibonacci
 
 class TestEvenFibonacci(unittest.TestCase):
     def test_throws_if_bad_input(self):
-        fibonacci = implementation.Fibonacci()
+        fibonacci = Fibonacci()
         with self.assertRaises(Exception):
             fibonacci.get_even_sum(-1)
 
     def test_does_not_throw_if_correct_input(self):
-        fibonacci = implementation.Fibonacci()
+        fibonacci = Fibonacci()
         fibonacci.get_even_sum(1)
 
 if __name__ == '__main__':
